chore(cdm): split service responsibilities
